Delete extra files
