--- client/src/components/App.js
@@ -1,21 +1,0 @@
-import React from 'react'
-import SignUp from "./SignUp"
-import SignIn from "./SignIn"
-import background from "../images/snowRoad.jpg"
-
-function App() {
-    return (
-        <div  >
-
-
-
-                <div >
-                <SignUp />
-                </div>
-                 <SignIn />
-                </div>
-    
-    )
-}
-
-export default App
